Add color utility functions and fix font helper
- Add get_opposite(), comp_min(), comp_max(), is_empty() color utilities
- Add to_hsl(), to_rgb(), get_highlighted() HSL conversion functions
- Fix EPSILON macro conflict (renamed to COLOR_EPSILON)
- Fix float conversion warning in measure_text_internal()
- All functions are backward compatible - no existing API changes

+++ src/font_helper.h
@@ -112,7 +112,8 @@ inline raylib::Font load_font_for_string(const std::string &content,
 }
 
 inline float measure_text_internal(const char *content, const float size) {
-    return raylib::MeasureText(content, size);
+    return static_cast<float>(
+        raylib::MeasureText(content, static_cast<int>(size)));
 }
 inline vec2 measure_text(const raylib::Font font, const char *content,
                          const float size, const float spacing) {
